Latest Notebooks used for Colorado Regional
These are the latest Jupyter Notebooks,  input files it was expecting and output files it generated that was used at the Colorado Regional to process the JSON files exported from the Robot Scouter phone app.  The scouts were using the PIT template on Thursday's Practice Day and the Match Metrics template on Friday and Saturday.